Move pipeline cache stuff to its own module
